requirements.txt
numpy
pandas
scipy
scikit-learn
joblib
pyserial
pynput
matplotlib
